Move TestingSystemSecurityMetadata to trino-main
Currently it's used only for tests of `SystemSecurityMetadata`, but it
can be useful for way more than that.
