+++ testing/trino-tests/src/test/java/io/trino/server/testing/TestSystemSecurityMetadata.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.security;
+package io.trino.server.testing;
 
 import com.google.common.collect.ImmutableSet;
 import io.trino.Session;
